#49: apply clean architecture layers
