+++ src/main/java/dev/orion/users/data/usecases/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Bussines rules package.
+ */
+package dev.orion.users.data.usecases;
