+++ src/main/java/dev/orion/users/domain/dto/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Data transfer objects packages.
+ */
+package dev.orion.users.domain.dto;
