+++ src/main/java/dev/orion/users/domain/model/Role.java
@@ -1,0 +1,47 @@
+/**
+ * @License
+ * Copyright 2022 Orion Services @ https://github.com/orion-services
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.orion.users.domain.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotNull;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
+import io.quarkus.hibernate.reactive.panache.PanacheEntityBase;
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * Role Entity.
+ */
+@Entity
+@Getter
+@Setter
+public class Role extends PanacheEntityBase {
+
+    /** Primary key. */
+    @Id
+    @GeneratedValue
+    @JsonIgnore
+    private Long id;
+
+    /** The name of the user. */
+    @NotNull(message = "The name of the role can't be null")
+    private String name;
+}
