+++ src/main/java/dev/orion/users/domain/model/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Model package.
+ */
+package dev.orion.users.domain.model;
