--- src/main/java/dev/orion/users/dto/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Data transfer objects packages.
- */
-package dev.orion.users.dto;
