--- src/main/java/dev/orion/users/model/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Model package.
- */
-package dev.orion.users.model;
