+++ src/main/java/dev/orion/users/presentation/exceptions/UserWSException.java
@@ -1,0 +1,58 @@
+/**
+ * @License
+ * Copyright 2022 Orion Services @ https://github.com/orion-services
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.orion.users.presentation.exceptions;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response.Status;
+
+/**
+ * Service exception.
+ */
+public class UserWSException extends WebApplicationException {
+
+    /**
+     * Service Exception constructor.
+     *
+     * @param message : The message of the exception
+     * @param status  : The HTTP error code
+     */
+    public UserWSException(final String message, final Status status) {
+        super(init(message, status));
+    }
+
+    /**
+     * A static method to init the message.
+     *
+     * @param message : An error message
+     * @param status  : A HTTP error code
+     *
+     * @return A Response object
+     */
+    private static Response init(final String message, final Status status) {
+        List<Map<String, String>> violations = new ArrayList<>();
+        violations.add(Map.of("message", message));
+        return Response.status(status)
+                .entity(Map.of("violations", violations))
+                .build();
+    }
+
+}
