+++ src/main/java/dev/orion/users/presentation/exceptions/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Web service exceptions.
+ */
+package dev.orion.users.presentation.exceptions;
