+++ src/main/java/dev/orion/users/presentation/mail/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * E-mail resources.
+ */
+package dev.orion.users.presentation.mail;
