+++ src/main/java/dev/orion/users/presentation/services/authentication/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Authentication WS.
+ */
+package dev.orion.users.presentation.services.authentication;
