+++ src/main/java/dev/orion/users/presentation/services/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Web services package.
+ */
+package dev.orion.users.presentation.services;
