--- src/main/java/dev/orion/users/usecase/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Bussines rules package.
- */
-package dev.orion.users.usecase;
