--- src/main/java/dev/orion/users/ws/authentication/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Authentication WS.
- */
-package dev.orion.users.ws.authentication;
