--- src/main/java/dev/orion/users/ws/exceptions/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Web service exceptions.
- */
-package dev.orion.users.ws.exceptions;
