--- src/main/java/dev/orion/users/ws/mail/package-info.java
@@ -1,4 +1,0 @@
-/**
- * E-mail resources.
- */
-package dev.orion.users.ws.mail;
