--- src/main/java/dev/orion/users/ws/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Web services package.
- */
-package dev.orion.users.ws;
