+++ src/test/java/dev/orion/users/TwoFactorAuthHandlerUnitTest.java
@@ -19,7 +19,7 @@ import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 import com.google.zxing.WriterException;
 
-import dev.orion.users.ws.handlers.TwoFactorAuthHandler;
+import dev.orion.users.presentation.handlers.TwoFactorAuthHandler;
 
 @ExtendWith(MockitoExtension.class)
 @TestMethodOrder(OrderAnnotation.class)
